pipeline/azure_deployer.py
"""
Azure deployment orchestrator for bus transit simulation pipeline.
"""

import os
import yaml
import json
import time
import logging
from datetime import datetime, timedelta
from typing import Dict, Any, Optional, List

try:
    from azure.identity import DefaultAzureCredential
    from azure.mgmt.containerinstance import ContainerInstanceManagementClient
    from azure.storage.blob import BlobServiceClient
    from azure.mgmt.resource import ResourceManagementClient
    AZURE_AVAILABLE = True
except ImportError:
    AZURE_AVAILABLE = False
    print("Azure SDK not available. Install with: pip install azure-identity azure-mgmt-containerinstance azure-storage-blob azure-mgmt-resource")


class AzureDeployer:
    """
    Handles deployment of bus simulation pipeline to Azure services.
    """
    
    def __init__(self, config_path: str = "config/azure_config.yaml"):
        """Initialize the Azure deployer."""
        if not AZURE_AVAILABLE:
            raise ImportError("Azure SDK not available. Please install required packages.")
            
        self.config_path = config_path
        self.config = self._load_config()
        self.logger = self._setup_logging()
        
        # Initialize Azure clients
        self.credential = DefaultAzureCredential()
        self.subscription_id = os.getenv('AZURE_SUBSCRIPTION_ID')
        
        if not self.subscription_id:
            raise ValueError("AZURE_SUBSCRIPTION_ID environment variable must be set")
    
    def _load_config(self) -> Dict[str, Any]:
        """Load Azure configuration from YAML file."""
        try:
            with open(self.config_path, 'r') as file:
                return yaml.safe_load(file)
        except Exception as e:
            print(f"Error loading Azure config: {str(e)}")
            return {}
    
    def _setup_logging(self) -> logging.Logger:
        """Set up logging for the deployer."""
        logger = logging.getLogger('AzureDeployer')
        logger.setLevel(logging.INFO)
        
        if not logger.handlers:
            handler = logging.StreamHandler()
            formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
            handler.setFormatter(formatter)
            logger.addHandler(handler)
        
        return logger
    
    def deploy_container_instance(self, environment: str = "development") -> Optional[str]:
        """Deploy simulation as Azure Container Instance."""
        self.logger.info(f"Deploying to Azure Container Instances - {environment}")
        # Implementation would go here
        return f"bus-simulation-{environment}-{int(time.time())}"


def main():
    """Main function for testing the Azure deployer."""
    deployer = AzureDeployer()
    container_name = deployer.deploy_container_instance("development")
    print(f"Deployed container: {container_name}")


if __name__ == "__main__":
    main() 