requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0

# Configuration management
PyYAML>=6.0

# Azure SDK for cloud integration
azure-storage-blob>=12.17.0
azure-identity>=1.13.0
azure-keyvault-secrets>=4.7.0
azure-monitor-opentelemetry-exporter>=1.0.0

# Logging and monitoring
opencensus-ext-azure>=1.1.0

# Progress tracking
tqdm>=4.65.0

# Date/time utilities
python-dateutil>=2.8.2

# Optional: For advanced analytics and visualization
# matplotlib>=3.7.0
# seaborn>=0.12.0
# plotly>=5.14.0

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0 